Updated EA Query

+++ include/nigiri/tripbased/trip_segment.h
@@ -10,7 +10,7 @@ struct trip_segment {
   trip_segment(transport_idx_t t_idx,
                size_t from_idx,
                size_t to_idx,
-               size_t n_transfers,
+               uint8_t n_transfers,
                bool on_query_day)
       : t_idx_(t_idx),
         from_idx_(from_idx),
@@ -24,7 +24,7 @@ struct trip_segment {
   trip_segment(transport_idx_t t_idx,
                size_t from_idx,
                size_t to_idx,
-               size_t n_transfers,
+               uint8_t n_transfers,
                size_t prev_idx,
                bool on_query_day)
       : t_idx_(t_idx),
@@ -38,7 +38,7 @@ struct trip_segment {
   size_t from() const { return from_idx_; };
   size_t to() const { return to_idx_; }
   size_t prev_idx() const { return prev_idx_; }
-  size_t n_transfers() const { return n_transfers_; }
+  uint8_t n_transfers() const { return n_transfers_; }
   bool on_query_day() const { return on_query_day_; }
 
 private:
@@ -46,7 +46,7 @@ private:
   size_t from_idx_;
   size_t to_idx_;
   size_t prev_idx_;
-  size_t n_transfers_;
+  uint8_t n_transfers_;
   bool on_query_day_;
 };
 }  // namespace nigiri::tripbased
